Refactor storage into sub package
storage/page_blob.py
storage/container.py
storage/storage.py
