+++ azurectl/account/storage.py
@@ -13,7 +13,7 @@
 #
 # project
 from ..service_manager import ServiceManager
-from ..container import Container
+from ..storage.container import Container
 from ..defaults import Defaults
 from ..azurectl_exceptions import (
     AzureStorageAccountCreateError,
