+++ azurectl/storage/fileshare.py
@@ -14,7 +14,7 @@
 from azure.storage.file import FileService
 
 # project
-from azurectl_exceptions import (
+from ..azurectl_exceptions import (
     AzureFileShareListError,
     AzureFileShareCreateError,
     AzureFileShareDeleteError
