+++ azurectl/storage/page_blob.py
@@ -11,7 +11,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-from azurectl_exceptions import (
+from ..azurectl_exceptions import (
     AzurePageBlobZeroPageError,
     AzurePageBlobAlignmentViolation,
     AzurePageBlobSetupError,
